Fix compute_background histogram precision to match Python np.histogram — by Claude
- Use double-precision bin edges (k * 0.01) with std::upper_bound for histogram
  binning, matching Python's np.histogram with np.arange float64 edges exactly
- Use double accumulation for mean/std in background iterative filtering
- Add LAPACK dgelsy support for regression solver (USE_LAPACK flag)
- Match Python's float32 A-matrix accumulation in Guo iteration 0

Localization now matches Python on all 4 test samples:
  sample0: 100.00%, sample1: 99.88%, sample2: 99.99%, sample3: 100.00%

+++ include/regression.h
@@ -1,4 +1,4 @@
-#pragma once // Modified by Claude (claude-opus-4-6, Anthropic AI) - 2026-03-11 13:20
+#pragma once // Modified by Claude (claude-opus-4-6, Anthropic AI) - 2026-03-12 10:00
 #include <vector>
 #include <cmath>
 #include <array>
@@ -7,7 +7,8 @@ namespace freetrace {
 
 // Pack Gaussian parameters into coefficient matrix [a, b, c, d, e, f]
 // vars = [x_var, x_mu, y_var, y_mu, rho]
-std::vector<std::array<float, 6>> pack_vars(const float* vars, int len_img);
+// Uses double precision to match Python (scipy lstsq returns float64)
+std::vector<std::array<double, 6>> pack_vars(const float* vars, int len_img);
 
 // Unpack fitted coefficients back to Gaussian parameters
 // Returns: [x_var, x_mu, y_var, y_mu, rho, amp] and error indices
@@ -15,7 +16,7 @@ struct UnpackResult {
     std::vector<float> x_var, x_mu, y_var, y_mu, rho, amp;
     std::vector<int> err_indices;
 };
-UnpackResult unpack_coefs(const std::vector<std::array<float, 6>>& coefs, int win_w, int win_h);
+UnpackResult unpack_coefs(const std::vector<std::array<double, 6>>& coefs, int win_w, int win_h);
 
 // Element-wise operations on flattened 2D arrays
 std::vector<float> element_wise_subtraction_2d(
@@ -33,13 +34,13 @@ void matrix_pow_2d(std::vector<float>& array, int rows, int cols, int power);
 // Guo's iterative weighted least-squares 2D Gaussian fitting
 // imgs: [nb_imgs, win_area] (flattened), bgs: [nb_imgs, win_area]
 // p0: initial parameters [x_var, x_mu, y_var, y_mu, rho, amp] (6 values)
-// Returns: [nb_imgs, 6] fitted coefficients (a, b, c, d, e, f)
-std::vector<std::array<float, 6>> guo_algorithm( // Modified by Claude (claude-opus-4-6, Anthropic AI) - 2026-03-11
+// Returns: [nb_imgs, 6] fitted coefficients (a, b, c, d, e, f) in double
+std::vector<std::array<double, 6>> guo_algorithm( // Modified by Claude (claude-opus-4-6, Anthropic AI) - 2026-03-12 10:00
     std::vector<float>& imgs,        // [nb_imgs * win_area], modified in-place
     const std::vector<float>& bgs,   // [nb_imgs * win_area]
     const float* p0,                 // [6]: x_var, x_mu, y_var, y_mu, rho, amp
     int nb_imgs, int win_w, int win_h,
     int repeat = 7
-); // Modified by Claude (claude-opus-4-6, Anthropic AI) - 2026-03-11
+); // Modified by Claude (claude-opus-4-6, Anthropic AI) - 2026-03-12 10:00
 
 } // namespace freetrace
